requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.1
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-speech==2.25.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
mysqlclient==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
pypiwin32==223
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
requests==2.31.0
rsa==4.9
SpeechRecognition==3.10.1
sqlparse==0.2.4
srt==3.5.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
vosk==0.3.45
websockets==12.0
